+++ integration/app/list/list.component.ts
@@ -19,8 +19,8 @@ export class ListComponent {
   hello$ = this._store.select(ListState.getHello);
   hello = this._store.selectSignal(ListState.getHello);
 
-  router$ = this._store.select<RouterStateSnapshot | undefined>(RouterState.state);
-  router = this._store.selectSignal<RouterStateSnapshot | undefined>(RouterState.state);
+  router$ = this._store.select<RouterStateSnapshot | undefined>(RouterState.state());
+  router = this._store.selectSignal<RouterStateSnapshot | undefined>(RouterState.state());
 
   constructor(private _store: Store) {}
 }