+++ packages/router-plugin/tests/router-data-resolved.spec.ts
@@ -55,7 +55,7 @@ describe('RouterDataResolved', () => {
 
     constructor(store: Store) {
       this.router$ = store.select(
-        RouterState.state
+        RouterState.state()
       ) as unknown as Observable<RouterStateModel>;
     }
   }
@@ -110,7 +110,7 @@ describe('RouterDataResolved', () => {
       expect(dataFromTheOriginalRouter).toEqual({ test });
 
       const dataFromTheRouterState = store.selectSnapshot<RouterStateSnapshot | undefined>(
-        RouterState.state
+        RouterState.state()
       )!.root.firstChild!.data;
       expect(dataFromTheOriginalRouter).toEqual(dataFromTheRouterState);
     })
@@ -160,7 +160,7 @@ describe('RouterDataResolved', () => {
       expect(dataFromTheOriginalRouter).toEqual({ test });
 
       const dataFromTheRouterState = store.selectSnapshot<RouterStateSnapshot | undefined>(
-        RouterState.state
+        RouterState.state()
       )!.root.firstChild!.data;
       expect(dataFromTheOriginalRouter).toEqual(dataFromTheRouterState);
     })